add sdp sanitize
